Crate read_color

A simple library for reading hex colors.

Functions

char_to_hex
Converts a character into a u8 value.

hex
Reads a hex value from an iterator of characters.

hex_pair
Reads a pair of hex values, joining them in value range 0-255.

rgb
Reads RGB colors from iterator of characters.

rgb_maybe_a
Reads RGB with optional alpha from iterator of characters.

rgba
Reads RGBA colors from iterator of characters.